move equipment canonicalization to Equipment.ts, call it in calculateEquipmentBonusesFromGear
